+++ fastly/domainmanagement/doc.go
@@ -1,0 +1,3 @@
+// Package domainmanagement contains subpackages which offer various operations to
+// enable, disable, and configure Fastly domains
+package domainmanagement