+++ fastly/domainmanagement/v1/domains/api_update.go
@@ -1,4 +1,4 @@
-package v1
+package domains
 
 import (
 	"context"
@@ -26,7 +26,7 @@ func Update(ctx context.Context, c *fastly.Client, i *UpdateInput) (*Data, error
 		return nil, fastly.ErrMissingDomainID
 	}
 
-	path := fastly.ToSafeURL("domains", "v1", *i.DomainID)
+	path := fastly.ToSafeURL("domain-management", "v1", "domains", *i.DomainID)
 
 	resp, err := c.PatchJSON(ctx, path, i, fastly.CreateRequestOptions())
 	if err != nil {