+++ fastly/domains/doc.go
@@ -1,3 +1,0 @@
-// Package domains contains subpackages which offer various operations to
-// enable, disable, and configure Fastly domains
-package domains